Only start checkpoint cleanup thread on the leader
Gate the initial thread launch in __wt_checkpoint_cleanup_create on the
disaggregated role: a follower opens with the thread stopped, and a
subsequent step-up starts it. Non-disaggregated connections still start
the thread unconditionally at open. Rename the role-transition test to
test_layered_checkpoint_cleanup01 and cover both follower-open and
leader-open startup paths with absolute counter assertions.
